# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Foodie app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of foodie's app"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/recipes": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get all recipes",
        "description": "This endpoint retrieves a list of all recipes based on the provided filters such as owner, category, area, ingredient, and favorite.\nThe response will return a paginated list of recipes along with their details including the owner name and avatar.\n",
        "operationId": "getAllRecipes",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "The page number to fetch (pagination)",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "The number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 20
            }
          },
          {
            "in": "query",
            "name": "favorite",
            "description": "The user ID to filter recipes that are marked as favorites by the user",
            "required": false,
            "schema": {
              "type": "string",
              "example": "user Id"
            }
          },
          {
            "in": "query",
            "name": "category",
            "description": "The category ID to filter recipes",
            "required": false,
            "schema": {
              "type": "string",
              "example": "some-category-id"
            }
          },
          {
            "in": "query",
            "name": "ingredient",
            "description": "The ingredient ID to filter recipes that contain a specific ingredient",
            "required": false,
            "schema": {
              "type": "string",
              "example": "some-ingredient-id"
            }
          },
          {
            "in": "query",
            "name": "area",
            "description": "The area ID to filter recipes by area",
            "required": false,
            "schema": {
              "type": "string",
              "example": "some-area-id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all recipes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "recipes",
                    "page",
                    "limit",
                    "totalItem",
                    "totalPage",
                    "hasPreviousPage",
                    "hasNextPage"
                  ],
                  "properties": {
                    "recipes": {
                      "type": "array",
                      "description": "List of retrieved recipes",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "title",
                          "categoryId",
                          "ownerId",
                          "areaId",
                          "description",
                          "time",
                          "thumb",
                          "ownerName",
                          "ownerAvatar",
                          "isFavorite"
                        ],
                        "properties": {
                          "id": {
                            "description": "Unique identifier for the recipe",
                            "type": "string",
                            "example": "recipe id"
                          },
                          "title": {
                            "description": "Title of the recipe",
                            "type": "string",
                            "example": "Battenberg Cake"
                          },
                          "categoryId": {
                            "description": "Category ID the recipe belongs to",
                            "type": "string",
                            "example": "category id"
                          },
                          "ownerId": {
                            "description": "ID of the user who owns the recipe",
                            "type": "string",
                            "example": "owner id"
                          },
                          "areaId": {
                            "description": "Area ID the recipe belongs to",
                            "type": "string",
                            "example": "area id"
                          },
                          "description": {
                            "description": "Brief description of the recipe",
                            "type": "string",
                            "example": "A classic British cake made with almond sponge cake and covered with marzipan."
                          },
                          "time": {
                            "description": "Time (in minutes) required to prepare the recipe",
                            "type": "integer",
                            "example": 60
                          },
                          "thumb": {
                            "description": "Thumbnail image URL for the recipe",
                            "type": "string",
                            "example": "https://example.com/recipe.jpg"
                          },
                          "ownerName": {
                            "description": "Name of the recipe owner",
                            "type": "string",
                            "example": "Owner name"
                          },
                          "ownerAvatar": {
                            "description": "Owner's avatar URL or null",
                            "oneOf": [
                              {
                                "type": "string",
                                "example": "https://example.com/avatar.jpg"
                              },
                              {
                                "type": "null",
                                "example": null
                              }
                            ]
                          },
                          "isFavorite": {
                            "description": "Indicates if the recipe is marked as favorite by the user",
                            "type": "boolean",
                            "example": false
                          }
                        }
                      }
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "description": "Number of recipes per page",
                      "example": 20
                    },
                    "totalItem": {
                      "type": "integer",
                      "description": "Total number of recipes available",
                      "example": 1
                    },
                    "totalPage": {
                      "type": "integer",
                      "description": "Total number of pages available",
                      "example": 1
                    },
                    "hasPreviousPage": {
                      "type": "boolean",
                      "description": "Indicates if there is a previous page",
                      "example": false
                    },
                    "hasNextPage": {
                      "type": "boolean",
                      "description": "Indicates if there is a next page",
                      "example": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Recipes"
        ],
        "summary": "Create a new recipe with thumbnail",
        "description": "This endpoint allows authenticated users to create new recipes.\nIngredients should be sent as a JSON string in the 'ingredients' field.\n",
        "operationId": "createRecipe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "time",
                  "categoryId",
                  "ingredients"
                ],
                "properties": {
                  "thumb": {
                    "type": "string",
                    "format": "binary",
                    "description": "Recipe thumbnail image (optional)"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 128,
                    "example": "Grilled Squid"
                  },
                  "time": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 30
                  },
                  "instructions": {
                    "type": "string",
                    "example": "Step 1 Prepare the squid..."
                  },
                  "description": {
                    "type": "string",
                    "example": "A delicious grilled squid recipe"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "6462a6cd4c3d0ddd28897f95"
                  },
                  "areaId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "6462a6f04c3d0ddd28897f9e"
                  },
                  "ingredients": {
                    "type": "string",
                    "description": "JSON array of ingredients as string.\nExample: [{\"ingredientId\":\"640c2dd963a319ea671e3664\",\"measure\":\"100 g\"}]\n",
                    "example": "[{\"ingredientId\":\"640c2dd963a319ea671e3664\",\"measure\":\"100 g\"}]"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "new recipe id"
                    },
                    "title": {
                      "type": "string",
                      "example": "Grilled Squid"
                    },
                    "categoryId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "category id"
                    },
                    "ownerId": {
                      "type": "string",
                      "example": "recipe owner id"
                    },
                    "areaId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "area id"
                    },
                    "instructions": {
                      "type": "string",
                      "example": "Step 1 Prepare the squid..."
                    },
                    "description": {
                      "type": "string",
                      "example": "A delicious grilled squid recipe"
                    },
                    "time": {
                      "type": "integer",
                      "example": 1
                    },
                    "thumb": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example/recipe.jpg"
                    },
                    "ingredients": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "ingredient id"
                          },
                          "name": {
                            "type": "string",
                            "example": "Bacon"
                          },
                          "measure": {
                            "type": "string",
                            "example": "100 g"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Possible reasons:\n- Invalid input data\n- Missing required fields\n- Invalid ingredients format\n"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/recipes/popular": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get popular recipes",
        "description": "This endpoint retrieves a list of the most popular recipes based on the number of times they have been favorited by users.\nThe response includes recipe details, owner information, and whether the recipe is favorited by the active user.\n",
        "operationId": "getPopularRecipes",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "description": "The number of recipes to fetch (default is 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved popular recipes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "recipeId",
                      "recipeName",
                      "recipeDescription",
                      "recipeUserName",
                      "recipeUserAvatar",
                      "recipeLikeStatus"
                    ],
                    "properties": {
                      "recipeId": {
                        "description": "Unique identifier for the recipe",
                        "type": "string",
                        "example": "recipe-id"
                      },
                      "recipeName": {
                        "description": "Title of the recipe",
                        "type": "string",
                        "example": "Grilled Squid"
                      },
                      "recipeDescription": {
                        "description": "Brief description of the recipe",
                        "type": "string",
                        "example": "A delicious grilled squid recipe for seafood lovers."
                      },
                      "recipeUserName": {
                        "description": "Name of the user who owns the recipe",
                        "type": "string",
                        "example": "John Doe"
                      },
                      "recipeUserAvatar": {
                        "description": "Avatar URL of the user who owns the recipe",
                        "type": "string",
                        "example": "https://example.com/avatar.jpg"
                      },
                      "recipeLikeStatus": {
                        "description": "Indicates whether the recipe is favorited by the active user",
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/recipes/{id}/own-recipes": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get user's own recipes with pagination",
        "description": "Returns a paginated list of recipes created by the specified user.\nIncludes ingredients list and favorite status for authenticated users.\n",
        "operationId": "getUserOwnRecipes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the user whose recipes to retrieve"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recipes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "recipe id"
                          },
                          "title": {
                            "type": "string",
                            "example": "Recipe title"
                          },
                          "description": {
                            "type": "string",
                            "example": "Recipe description"
                          },
                          "time": {
                            "type": "integer",
                            "example": 30
                          },
                          "thumb": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://example.com/image.jpg"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-05-15T10:00:00.000Z"
                          },
                          "ownerName": {
                            "type": "string",
                            "example": "Owner name"
                          },
                          "ownerAvatar": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://example.com/avatar.jpg"
                          },
                          "ingredients": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "example": "ingredient id"
                                },
                                "name": {
                                  "type": "string",
                                  "example": "Bacon"
                                },
                                "img": {
                                  "type": "string",
                                  "format": "uri",
                                  "example": "https://example.com/ingredient.jpg"
                                },
                                "measure": {
                                  "type": "string",
                                  "example": "200 г"
                                }
                              }
                            }
                          },
                          "isFavorite": {
                            "type": "boolean",
                            "example": true
                          }
                        }
                      }
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 5
                    },
                    "totalRecipes": {
                      "type": "integer",
                      "example": 100
                    },
                    "recipesPerPage": {
                      "type": "integer",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Possible reasons:\n- Invalid user ID format\n- Page number out of range\n"
          },
          "401": {
            "description": "Unauthorized (missing or invalid token)"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/{id}/favorite": {
      "patch": {
        "tags": [
          "Recipes"
        ],
        "summary": "Update favorite status of a recipe",
        "description": "This endpoint allows an authenticated user to mark or unmark a recipe as favorite.\nThe request body must include the `favorite` field (boolean).\nThe response will return the recipe ID and the updated favorite status.\n",
        "operationId": "updateFavoriteController",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the recipe to update favorite status",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6462a8f74c3d0ddd28897fcd"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "favorite"
                ],
                "properties": {
                  "favorite": {
                    "description": "Indicates whether the recipe is marked as favorite",
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the favorite status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "recipeId",
                    "favorite"
                  ],
                  "properties": {
                    "recipeId": {
                      "description": "ID of the recipe",
                      "type": "string",
                      "example": "6462a8f74c3d0ddd28897fcd"
                    },
                    "favorite": {
                      "description": "Updated favorite status of the recipe",
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input data)"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Not authorized"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                }
              }
            }
          }
        }
      }
    }
  }
}